backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
slowapi==0.1.9
cryptography==42.0.5
google-generativeai>=0.7.2
groq==0.9.0
openai==1.30.1
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
aiofiles==23.2.1
anthropic==0.28.0
fal-client==0.4.0
